Fixed an issue with // being generated for some invite links.
Removed accessibile API as it's no longer needed.
